Revert "Teleop Changes"
