+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/subsystems/Carousel.java
@@ -5,6 +5,7 @@ import com.qualcomm.robotcore.hardware.HardwareMap;
 
 public class Carousel {
 
+
     private DcMotor carouselMotor; //Port
 
     public void init(HardwareMap hardwareMap) {
@@ -22,6 +23,7 @@ public class Carousel {
             carouselMotor.setPower(0);
         }
     }
+
 
     public DcMotor getCarouselMotor() {
         return carouselMotor;
